backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
anthropic>=0.28.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-anthropic>=0.1.0
pymupdf>=1.24.0
Pillow>=10.0.0
filetype>=1.2.0
reportlab>=4.0.0
aiofiles>=23.0.0
httpx>=0.27.0
